Split test case into separate class. Require sequential execution.
