+++ asciidoctorj-core/src/test/java/org/asciidoctor/WhenAsciidoctorLogsToConsole.java
@@ -255,16 +255,6 @@ public class WhenAsciidoctorLogsToConsole {
         }
     }
 
-    @Test
-    public void shouldCreateNewLogHandlersOnCreateAsciidoctor() {
-        TestLogHandlerService.instancesCount.set(0);
-        Asciidoctor asciidoctor = Asciidoctor.Factory.create();
-        asciidoctor.convert("Test", Options.builder().build());
-        Asciidoctor asciidoctor2 = Asciidoctor.Factory.create();
-        asciidoctor2.convert("Test", Options.builder().build());
-        assertEquals(2, TestLogHandlerService.instancesCount.get());
-    }
-
     @Name("big")
     public static class LoggingProcessor extends BlockProcessor {
